pyproject.toml
[tool.poetry]
name = "codesearch-backend"
version = "0.1.0"
description = ""
authors = ["John"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.13.2"
fastapi = "^0.114.1"
sqlalchemy = "^2.0.34"
uvicorn = "^0.30.6"
gitpython = "^3.1.43"
pydantic = {extras = ["email"], version = "^2.9.1"}
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
black = "^24.8.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.14.0"}
anthropic = "0.32.0"
infomap = "2.8.0"
intervaltree = "3.1.0"
llama-index = "0.10.59"
networkx = "3.3"
openai = "1.38.0"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
setuptools = "65.5.0"
simple-parsing = "0.1.5"
starlette = "0.38.2"
tenacity = "8.5.0"
tiktoken = "0.7.0"
tree-sitter = "0.22.3"
tree-sitter-java = "0.21.0"
tree-sitter-python = "0.21.0"
typing-extensions = "4.12.2"
bcrypt = "^4.2.0"
boto3 = "^1.35.19"
rapidfuzz = "^3.9.7"
faiss-cpu = "^1.8.0.post1"
astroid = "~3"
pylint = "^3.2.7"
pytest = "^8.3.3"
instructor = "~1"
sqlalchemy-utils = "^0.41.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"